Remove platform build specs from archived conda environment
By default the `conda env export` includes the platform specific build
strings in the specified versions, but that means the environment can't
be used on any platform other than the one it was created on, which
isn't what we're going for.

Also, filter out the unused (and confusing) "prefix" line from the
archived environment file.